:lock: Send along the webhook ID in HTTP headers
This way, we can send the same webhook multiple times, e.g. in case of
failure, and have the receiver know it's just a resend of the same
webhook (because they already saw that id). If they have properly
implemented their endpoints, we can safely retry our webhooks when
something seemed wrong the first time.

closes #1